+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## [0.10.0-nullsafety.1] - beta version:
+* improve phone text field
+* recuperate calling code with phone number
+* show country flag by calling code
+* country calling code picker
 ## [0.9.0-nullsafety.1] - beta version:
 * fix readme badge
 ## [0.9.0-nullsafety.0] - beta version: